Moving find_leaves and find_roots to core.py since it's needed for graph traversal
